chore(toggle-input): move to packages
