+++ packages/components/inputs/toggle-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './toggle-input';
+
+export { default as version } from './version';
