+++ packages/components/inputs/toggle-input/src/toggle.story.js
@@ -3,8 +3,8 @@ import { Value } from 'react-value';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, boolean, select, text } from '@storybook/addon-knobs/react';
 import { action } from '@storybook/addon-actions';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import Toggle from '.';
 
 storiesOf('Components|Inputs', module)
