--- src/components/inputs/toggle-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './toggle-input';
